Test Cases Finished
Test cases:
- Organizer sign up and create event
- Admin sign up and edit profile (first and last name)
- Attendee sign up
