+++ app/src/main/res/navigation/admin_nav_graph.xml
@@ -42,6 +42,6 @@
     </fragment>
     <fragment
         android:id="@+id/adminProfileFragment"
-        android:name="com.example.queueup.views.admin.AdminProfileFragment"
+        android:name="com.example.queueup.views.profiles.ProfileFragment"
         android:label="AdminProfileFragment" />
 </navigation>
